requirements.txt
# FastAPI и сервер
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.25
alembic==1.13.1

# Аутентификация
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Валидация данных
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Прогнозирование и данные
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# Утилиты
python-dotenv==1.0.0
pyyaml==6.0.1